- put fstrm protocol contents in separate files, dnstap_fstrm.c   and dnstap_fstrm.h
